making API: Authorisation and Authentication

+++ controllers/api/v1/post_api.js
@@ -1,5 +1,6 @@
 const Post = require("../../../models/post");
 const Comment = require("../../../models/comment");
+const User = require("../../../models/user");
 
 module.exports.index = async function (req, res) {
   let posts = await Post.find({})
@@ -21,18 +22,19 @@ module.exports.index = async function (req, res) {
 module.exports.destroy = async function (req, res) {
   try {
     let post = await Post.findById(req.params.id);
+    if (post.user == req.user.id) {
+      post.remove();
 
-    post.remove();
+      await Comment.deleteMany({ post: req.params.id });
 
-    await Comment.deleteMany({ post: req.params.id });
-
-    return res.status(200).json({
-      message: "Post and associated comments deleted successfully",
-    });
-
-    // } else {
-    //   return res.redirect("back");
-    // }
+      return res.status(200).json({
+        message: "Post and associated comments deleted successfully",
+      });
+    } else {
+      return res.status(401).json({
+        message: "You cannot delete this post!",
+      });
+    }
   } catch (err) {
     return res.status(500).json({
       message: "Internal Server Error",
